refactor: move validation and data dispatch to separate funcs
